Updates for JOBS file formatting
This handles opening files in the JOBS format that, for some reason, with the original
script open as separate series for every time point, and therefore you can only ever
segment the first frame.

Unfortunately this currently overflows the memory of FIJI because you can't do a virtual
stack with the concatenation strategy, so I will need to fix that soon.